* Removing X-modem code * Fixing issue with install.sh and images * Fixing issue where timer column on imu_t cannot accept large values of time.

+++ code/GroundStation/pgsql_tables.sql
@@ -14,7 +14,7 @@ CREATE TABLE imu_t (id integer primary key default nextval('imu_id_seq'),
                     accx      float,
                     accy      float,
                     accz      float,
-                    timer     integer,
+                    timer     bigint,
                     creation_date timestamp);
 
 ALTER TABLE imu_t ALTER COLUMN creation_date SET DEFAULT CURRENT_TIMESTAMP;
